+++ R/utils.R
@@ -8,7 +8,7 @@ str_trunc <- function(x, width) {
   str_width <- utf8::utf8_width(crayon::strip_style(x), encode = FALSE)
 
   too_wide <- which(!is.na(x) & str_width > width)
-  x[too_wide] <- paste0(crayon::col_substr(x[too_wide], 1, width - 1), get_ellipsis())
+  x[too_wide] <- paste0(fansi::substr_ctl(x[too_wide], 1, width - 1), get_ellipsis())
 
   x
 }